+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Bug Fixes:
     item type with an occurrence indicator and before the comma.
 1.  Fix parsing Saxon 9.8 `union` types with whitespace or comments after the
     QName and before the comma.
+1.  Fix parsing `TypedFunctionTest` constructs with whitespace or comments after the
+    item type with an occurrence indicator and before the comma.
 
 Inspections: